Optional master heartbeat via LastControllerPublish
At least with MPM3PM meter which only delivers Amps with 3 decimal digits it
seems that the heartbeat falsely triggers far too often.

With this commit, the LastControllerPublish MQTT message is the preferred
heartbeat value while current on phase with maximum current (the previous
implementation) is now the fallback in case LastControllerPublish message is
not published by master at all.