Bugfix: Actually do what debug printout says ...
If EV is not consuming the allowed power, algorithms might
calculate an allowed value that exceeds allowed system total current.
This is handled by  checks which work correct and print correct
debug messages. But they missed to do what the printout says.

In consquence, in some special cases, there could be osciallations of
allowed current in every control loop.. But as this only happened when
EV was anyway consuming by far less, it didn't affect the charge operation
at all. Still, it should be fixed.