Extend commandline parameters for external scripts
This commit adds three more parameters to the external calls to
on.py, off.py, and the generic watt.py: A name for an actor
(switch, relais, etc.), a username, and a password. The values
for these parameters are taken from the smarthomedevices configuration
in the same manner as for the device_ip.

For the sake of readability the construction of the list of arguments to
subprocess.Popen is done explicitely beforehand.