Share cache for same username and IP
To reduce the number of requests to FRITZ!Box devices, all Smarthome
Devices for this module that are configured with the same username and
the same Host/IP now share the cached session ID and device info. In the
worst case the cached values are 5 seconds old which is still fast
enough for OpenWB to work with.

In the unlikely event that a user mistyped one of the passwords, the
failed login session for a device might also inhibit value retrieval for
any other device that was supposed to use the same credentials. In case
of credential errors the log messages (in German language) will help to
identify the situationb.