EVU/WR Json Modules fix for complex Json Filters
To use complex JQ Filters wit selects and | (pipes) it is necessary to set the variables in double quotes so they are not interpreted in the shell.
Both modules are then usable for VZLogger to parse the JSON according to the UUIDs, so you need not rely on the Lines in VZLogger modules. In the JSON Module also the pvkwh is included which is missing in the vzlogger PV Module