Fix openwb.conf manipulations with sed when running in Docker
`sed -i` creates a temporary file and replaces the original file when
successful. This fails with "Device or resource busy" when openwb.conf
is mounted from a Docker host because it would change the inode.

This change replaces all invocations of `sed -i` on openwb.conf so that
the replacement result is stored in memory and then written to the
target file in a separate step.

See https://github.com/IngmarStein/openwb-docker/issues/3