Retry https://github.com/snaptec/openWB/pull/1404
The previous version clobbered the config file if the sed pattern
contained an error or sed failed for any reason. This attempt writes the
result into the target file only if the operation was successful,
storing the intermediate result in a temporary file.

This PR creates a new helper function `safeSed` to wrap this series of
commands.