Smarthome: Use SMA EnergyMeter as measuring module
I've reused the existing speedwiredecoder.py and created a module for using a SMA EnergyMeter for measuring.
My EnergyMeter is used for my heatpump and the heatpump is shut down twice a day. During this time the energy won't send and the current consumption is wrong. So I've implemented a workaround for handling an existing ret-file but changing the current consumption to 0 while the powerc remains.