Remove "content-length" from POST request headers
The Python module requests adds "content-length" headers automatically. No need to add them manually to the headers.
If you add them manually they are actually overwritten by mod requests. (And that is good because the length in the manual headers would be wrong anyway in most cases and would ruin the transmission.)